Some thoughts about the design of Timer
(Also, an example of resuming a paused timer that got me thinking about
the design.)

I don't like that the start/pause commands take the timer duration as a
parameter. The Timer should know how much time is left when it is paused
and it should know the configured timer duration when you start the
timer.

I designed it like this to mirror the Mobtime API, in which the start
and pause commands include the timer duration. I wanted the Timer class
to be able to receive commands both from the UI and from the Mobtime
server, without having to have two different interfaces for those
clients.