Remove unnecessary and broken optimization
The idea was to prevent Timer from spamming the Display. I made a
mistake in the implementation, which happened to not be a problem in the
test. Instead of checking the time since the last display update or
tick, it checks the time since the timer was started. This works
perfectly for the first second after the timer is started, but after
that it does nothing at all. Rather than fixing it, I decided to remove
the behavior and just rely on not ticking the Timer so much that it
becomes a problem. That's a nicer solution anyway.